Convert recruitment_banner_helper unit tests
- Move the fixture file used by this test
